Refactoring imports of TagSelection and StateSelection
